+++ test/polynomials/test_splines.jl
@@ -141,17 +141,7 @@ print("d_npass = $d_npass / 100: "); println_slim(@test d_npass > 95)
 
 @info("Testing `splinify`")
 
-function splinify(basis::RPIBasis)
-   basis1p = deepcopy(basis.pibasis.basis1p)
-   J_spl = ACE1.Splines.RadialSplines(basis1p.J; zlist = basis1p.zlist)
-   basis1p_spl = ACE1.RPI.BasicPSH1pBasis(J_spl, deepcopy(basis1p.SH), 
-                  deepcopy(basis1p.zlist), deepcopy(basis1p.spec), deepcopy(basis1p.Aindices) )
-   pibasis_spl = ACE1.PIBasis(basis1p_spl, deepcopy(basis1p.zlist), 
-               deepcopy(basis.pibasis.inner), deepcopy(basis.pibasis.evaluator))
-   basis_spl = RPIBasis(pibasis_spl, deepcopy(basis.A2Bmaps), deepcopy(basis.Bz0inds) )   
-   return basis_spl
-end
-
+import ACE1.Splines: splinify 
 basis_spl = splinify(rpibasis_P)
